+++ modules/ipc/src/zenoh/raw_publisher.cpp
@@ -25,9 +25,9 @@
 #include <zenoh_macros.h>
 
 #include "hephaestus/ipc/topic.h"
+#include "hephaestus/ipc/zenoh/conversions.h"
 #include "hephaestus/ipc/zenoh/service.h"
 #include "hephaestus/ipc/zenoh/session.h"
-#include "hephaestus/ipc/zenoh/utils.h"
 #include "hephaestus/serdes/type_info.h"
 #include "hephaestus/utils/exception.h"
